grug/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grug"
version = "0.1.0"
description = "Self-hosted AI agent for TTRPG Discord communities — core library"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "pydantic-ai-slim[anthropic,mcp]>=1.0.0",
    "anthropic>=0.80.0",
    "APScheduler>=3.10.0",
    "SQLAlchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=23.0.0",
    "httpx>=0.24.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

# Tell setuptools to look one level up (the repo root) for the `grug` package,
# since this pyproject.toml lives inside the `grug/` directory itself.
[tool.setuptools.packages.find]
where = [".."]
include = ["grug*"]
